Convert add fuel details to ints and floats within the add_fuel function

+++ utils.py
@@ -96,8 +96,8 @@ def calc_mpg(current_odometer, prev_odometer, litres_in):
     """
     Calculates the mpg figure for fuel entry
     """
-    mpg = ((float(current_odometer)
-            - float(prev_odometer)) / litres_in) * 4.544
+    mpg = ((current_odometer
+            - int(prev_odometer)) / litres_in) * 4.544
     return round(mpg, 2)
